Several source and header files added to implement UI functionalities.

+++ src/ui/core/glut.cpp
@@ -68,7 +68,7 @@ private:
    struct TimerInfo
    {
       TimerCallback* cb;
-      unsigned long  ms;
+      double         ms;
       double         timestamp;      
    };
